Move snmpsim fixture to conftest
