+++ tests/integration/mibs/sensors_test.py
@@ -1,9 +1,4 @@
 from itertools import cycle
-from shutil import which
-import subprocess
-import time
-import io
-import os
 
 import pytest
 import pytest_twisted
@@ -14,36 +9,6 @@ from nav.mibs import comet, pdu2_mib, powernet_mib
 
 
 ports = cycle(snmpprotocol.port() for i in range(50))
-
-
-@pytest.fixture(scope='session')
-def snmpsim():
-    snmpsimd = which('snmpsimd.py')
-    assert snmpsimd, "Could not find snmpsimd.py"
-    workspace = os.getenv('WORKSPACE', os.getenv('HOME', '/source'))
-    proc = subprocess.Popen(
-        [
-            snmpsimd,
-            '--data-dir={}/tests/integration/snmp_fixtures'.format(workspace),
-            '--log-level=error',
-            '--agent-udpv4-endpoint=127.0.0.1:1024',
-        ],
-        env={'HOME': workspace},
-    )
-
-    while not _lookfor('0100007F:0400', '/proc/net/udp'):
-        print("Still waiting for snmpsimd to listen for queries")
-        proc.poll()
-        time.sleep(0.1)
-
-    yield
-    proc.kill()
-
-
-def _lookfor(string, filename):
-    """Very simple grep-like function"""
-    data = io.open(filename, 'r', encoding='utf-8').read()
-    return string in data
 
 
 @pytest.fixture()
